fix: profile form refactor
